Add push notifications state and permissions to context
